7.1.3
7.1.3

Replaced QF_EVT_REF_CTR_INC_() with QEvt_refCtr_inc_() in ports
- embOS
- ESP-IDF
- Qt

7.1.3

7.1.3

7.1.3

7.1.3

7.1.3

7.1.3

7.1.2

first commit after fixing history

Revert "7.1.2"

This reverts commit 90cf4e1471b5e9c0853af97af8ec0bc67c7e19c6.

7.1.2

first commit after fixing the history

7.0.1

+++ examples/arm-cm/dpp_stm32f4-discovery/qxk/main.c
@@ -1,7 +1,7 @@
 /*****************************************************************************
 * Product: DPP example extened for QXK
-* Last updated for version 6.8.2
-* Last updated on  2020-07-15
+* Last updated for version 7.1.2
+* Last updated on  2022-10-05
 *
 *                    Q u a n t u m  L e a P s
 *                    ------------------------
@@ -79,7 +79,7 @@ int main() {
 
     /* start the extended thread */
     QXTHREAD_START(XT_Test1,                 /* Thread to start */
-                  (uint_fast8_t)1,           /* QP priority of the thread */
+                  1U,                        /* QP priority of the thread */
                   test1QueueSto,             /* message queue storage */
                   Q_DIM(test1QueueSto),      /* message length [events] */
                   test1StackSto,             /* stack storage */
@@ -91,7 +91,7 @@ int main() {
     /* start the Philo active objects... */
     for (n = 0U; n < N_PHILO; ++n) {
         QACTIVE_START(AO_Philo[n],           /* AO to start */
-                      (uint_fast8_t)(n + 3), /* QP priority of the AO */
+                      n + 3U,                /* QP priority of the AO */
                       philoQueueSto[n],      /* event queue storage */
                       Q_DIM(philoQueueSto[n]), /* queue length [events] */
                       (void *)0,             /* stack storage (not used) */
@@ -100,13 +100,13 @@ int main() {
     }
 
     /* example of prioritizing the Ticker0 active object */
-    QACTIVE_START(the_Ticker0, (uint_fast8_t)(N_PHILO + 3),
+    QACTIVE_START(the_Ticker0, N_PHILO + 3U,
                   0, 0, 0, 0, 0);
 
     /* NOTE: leave priority (N_PHILO + 4) free for mutex */
 
     QXTHREAD_START(XT_Test2,                 /* Thread to start */
-                  (uint_fast8_t)(N_PHILO + 5), /* QP priority of the thread */
+                  N_PHILO + 5U,              /* QP priority of the thread */
                   test2QueueSto,             /* message queue storage */
                   Q_DIM(test2QueueSto),      /* message length [events] */
                   test2StackSto,             /* stack storage */
@@ -116,7 +116,7 @@ int main() {
     /* NOTE: leave priority (N_PHILO + 6) free for mutex */
 
     QACTIVE_START(AO_Table,                  /* AO to start */
-                  (uint_fast8_t)(N_PHILO + 7), /* QP priority of the AO */
+                  N_PHILO + 7U,              /* QP priority of the AO */
                   tableQueueSto,             /* event queue storage */
                   Q_DIM(tableQueueSto),      /* queue length [events] */
                   (void *)0,                 /* stack storage (not used) */
